moved the lines_dataset and atom_data_f_dataset_not_prepared fixtures to test_atomic.py
